game: Fire a notification when the game finished
The architecture will be largely asynchronous and definitely not linear. Since a player can resign at any point of the game the match will notify external listeners about the fact that the game has come to an end.

Testing these kind of behaviour require a Test Double, because there are likely to be a number of different type that can listen to this event. Capturing this behaviour I use a mock library, called trompeloeil, which I had to add to the build system and hooked into the main of catch2. It would be nicer if the error message would be more like something an IDE can process but for now this will do.