refactor: Simplify the unit tests
The tests were getting too big for the implementation code they were testing. Remove the surplus testing (all listeners get notified is includes the case of a single listener).

Pulled up the `using namespace GameRules;` to the top level of the test module as it is used in every single test.