travis: Build missing packages in the conan install step
trompeloeil doesn't seem to have any binaries or at least nothing that is compatible with newer compilers. Adding the `--build=missing` flag to the conan install step will make sure that at least we try to build any missing dependency for the platform.

It is, however, suboptimal in that those binaries aren't supposed to be part of my own build at all, yet this way the CI system will keep building them.