game: Add basic simulation system
As an incremental step a unit now not only have a location but a destination. To reach the destination the update method will linearly interpolate between the current and the destination location.

For this to work at least semi-decently I had to change the Location to use floats. For now it will do, but probably this needs to be changed into some form of fixed point calculation later.