refactor(GameRules): Extract vector specific operations into its own type
The operations in the `Game::update(Duration)` functions are operating on a lower level of abstraction, when they reach for the coordinates. These operations are traditionally belong to a vector class.

As a design decision there is a conceptual difference between vectors and locations. The quantity of direction cannot be the mixed with the quantities of position on the map, they are fundamentally different. It is a basically the same idea as the having an index type and a fitting difference type.

Tranlation can be described by a vector, and it can translate a location type. Length, multiplication and divisionm, normalization all these functions make more sense on a vector than a location. Although the fields are named exactly the same (which is given by the coordinate system as such), their meaning is slightly different.