GameRules: Notify listeners of damage when a unit is under attack
Units can attack each other and when that happens a notification sent out to the interface about the damage.

The code has been changed significantly to accomodate the different type of commands that can be active for a given unit. Move, Attack and Idle commands are very different in that they take different set of arguments: Idle has none, Move has a simple location and Attack takes a target.

This change replaces the optionally present destination variable with the command object which is a variant. The idea is based on the std::variant example for FSM by Bartek's blog (https://www.bfilipek.com/2019/06/fsm-variant-game.html)

Further refinement will come to implement the concepts of the damage, the type of attack, and the attack radius.