GameRules: Implement attack radius
Having an attack radius means that unit can only inflict damage to other units if they reach the proximity of that unit from a specific distance.

As long as the unit executing an attack command is outside of this proximity they are going to approach their target, as if they were commanded to move.