refactor(GameRules): Remove the direct calls to update in tests
I think it wasn't a good decision to use a time in the Game::update method, so first I replace the occurences of the direct calls to this function in the tests, which will make it easier to remove the argument.

These tests really do not care about on what time resolution they work.