GameRules: Implement hit points and attack damage
When a unit is under attack, it suffers damage by loosing its hit points. Each time the unit is hit, it will removes the attack damage from its hit points.

This adds two more properties to the unit so it is time to use a more general approach to store the status of the unit.

Generally, the user of the GameRules.h should not be able to directly manipulate unit properties. It can set up when spawning a unit but after that, it is the rules of the game, the model in MVC terms, that controls the unit properties.

Used a builder pattern for preparing UnitProperties because it allows greater flexibility in initialisation of yet unknown properties.