build: add cmake option for UB/ASAN
Before the code is growing too big it's worth to turn on as many checks as possible. So far, at least on MacOS builds no warning and any undefined behaviour or address violation hit.

To keep it this way I would like to make sure I run the sanitizers for all the CI builds. It is not on by default because these sanitizers are basically instrumentated code which will much slower than should be.