testing: remove module tags for unit tests
There is no need for them because if you run the QuaRTS.UT with option `-#` it will use the file name base for as a tag so we can filter it by module.

The general design principle here is that a module is a single file, and so all related unit tests will go into that module.

Tags should be reserved for other purposes such as benchmarking, marking 'slow' tests for separate run, etc. Or perhaps collecting model, view and controller type of modules into their respective suite.