refactor: rename the GameRules namespace to game
The reason for it to simplify the naming convention. It also pays to have a distinct rule for namespaces as opposed to classes: Many modules will likely have a single main class and this way I can use the same name for the ns as to the class but different casing.

Namespaces are from now snail_cased.