refactor: remove the duration argument for `game::Game::update()`
The game is evaluated one step at the time and each time must be uniform in order to be deterministic. Instead the fundamental unit of time in the game will be a the update stage, a single 'tick'.

A single tick represents linear evolution within that time interval from one state to the next.