game: apply acceleration to units during attack command before they reach the attack radius of their targetunde
Acceleration is a universal attribute of a unit, should be applied whenever a unit is on the move somewhere. Until it reaches its prosribed maximum velocity it should accelerate. Units before they stop to inflict damage to their target under attack command, they essentially under a kind of move command and therefore the acceleration must be applied.